+++ CHANGELOG.md
@@ -12,6 +12,7 @@ This project adheres to [Semantic Versioning](http://semver.org/) and [Keep a Ch
 ### Fixes
 - No longer depends on the `pkg_resources` module being available from the setuptools package.
 
+- add usedforsecurity=False to md5 hash to indicate it is not used in security context
 
 ## v6.0.0 (2025-01-21)